requirements.txt
asgiref==3.5.0
Babel==2.10.1
Django==4.0.4
django-betterforms==1.2
django-money==2.1.1
django-phonenumber-field==6.1.0
phonenumbers==8.12.47
Pillow==9.1.0
py-moneyed==1.2
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
